python/requirements.txt
cv2
flask
flask_bootstrap
gpiozero
imutils
numpy
RPi.GPIO
visual-logging
wiringpi